+++ KeyboardKitDemoKeyboard/KeyboardViewController.swift
@@ -35,6 +35,7 @@ import KeyboardKit
 class KeyboardViewController: KeyboardInputViewController {
     
     
+    // MARK: - View Controller Lifecycle
     
     override func viewDidLoad() {
         super.viewDidLoad()